gs-overview-page: Add a distro-featured section
The distro-featured section shows applications, which are tagged
with "GnomeSoftware::DistroFeatured" key. This can be done with an appstream
data file saved into /usr/share/app-info/xmls/.  An example file can be found
at contrib/org.gnome.Software.DistroFeatured.xml . The applications can be
tagged with this key by other means, this is only one way to do it.

The section is shown if there are found enough applications (currently at least 12).